requirements.txt
pandas 
python-dotenv
numpy
setuptools
pymongo
pymongo[srv]==3.6
certifi
scikit-learn

# -e. this is needed only at the end of the project
